feat: redux version FSD 적용
