+++ poke-dex/src/features/Pokemon/ui/PokeCardList/PokeCardList.styles.ts
@@ -1,0 +1,18 @@
+import styled from "@emotion/styled";
+
+export const Loading = styled.div`
+  display: flex;
+  justify-content: center;
+`
+
+export const List = styled.ul`
+  list-style: none;
+  padding: 0;
+  margin: 0 0 32px 0;
+
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 20px;
+`;
