+++ poke-dex/src/features/Pokemon/ui/PokeCardList/PokeCardList.tsx
@@ -1,8 +1,9 @@
 import styled from "@emotion/styled";
-import PokeCard from "./PokeCard";
+import PokeCard from "../PokeCard/PokeCard";
 import { useEffect, useState } from "react";
-import { PokemonListResponseType, fetchPokemons } from "../Service/pokemonService";
+import { PokemonListResponseType, fetchPokemons } from "../../../../entities/Pokemon/api/pokemonService";
 import useInfiniteScroll from "react-infinite-scroll-hook";
+import { List, Loading } from "./PokeCardList.styles";
 
 const PokeCardList = () => {
   const [pokemons, setPokemons] = useState<PokemonListResponseType>({
@@ -51,22 +52,5 @@ const PokeCardList = () => {
     </>
   );
 };
-
-const Loading = styled.div`
-  display: flex;
-  justify-content: center;
-`
-
-const List = styled.ul`
-  list-style: none;
-  padding: 0;
-  margin: 0 0 32px 0;
-
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 20px;
-`;
 
 export default PokeCardList;
