+++ poke-dex/src/features/Pokemon/ui/PokeDetail/PokemonDetail.tsx
@@ -1,11 +1,11 @@
-import styled from "@emotion/styled";
-import PokeMarkChip from "../Common/PokeMarkChip";
-import { PokemonDetailType, fetchPokemonDetail } from "../Service/pokemonService";
+import PokeMarkChip from "../../../../shared/ui/PokeMarkChip/PokeMarkChip";
+import { PokemonDetailType, fetchPokemonDetail } from "../../../../entities/Pokemon/api/pokemonService";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import { PokeImageSkeletone } from "../Common/PokeImageSkeletone";
+import { PokeImageSkeletone } from "../../../../shared/ui/Icon/Icon";
 import { useSelector } from 'react-redux';
-import { RootState } from "../Store";
+import { RootState } from "../../../../shared/lib/Store";
+import { Body, Container, Divider, Footer, Image, ImageContainer, Table, TableHeader, TableRow } from "./PokemonDetail.styles";
 
 const PokemonDetail = () => {
   const { name } = useParams();
@@ -88,58 +88,5 @@ const PokemonDetail = () => {
     </Container>
   );
 }
-
-const Container = styled.section`
-  border: 1px solid #c0c0c0;
-  margin: 16px 32px;
-  border-radius: 16px;
-  box-shadow: 1px 1px 3px 1px #c0c0c0;
-`;
-const ImageContainer = styled.section`
-  display: flex;
-  flex: 1 1 auto;
-  justify-content: center;
-  align-items: center;
-  margin: 8px 0;
-  min-height: 350px;
-`;
-const Image = styled.img`
-  width: 350px;
-  height: 350px;
-`;
-const Divider = styled.hr`
-  margin: 32px;
-  border-styled: none;
-  border-top: 1px dashed #d3d3d3;
-`;
-const Body = styled.section`
-  margin: 0 32px;
-`;
-const Table = styled.table`
-  width: 100%;
-  border-collapse: collapse;
-  border-spacing: 0;
-  margin: 0 auto 16px;
-
-  th, td {
-    padding: 6px 10px;
-  }
-`;
-const TableRow = styled.tr`
-  border: 1px solid #f0f0f0;
-`;
-const TableHeader = styled.th`
-  width: 1px;
-  white-space: nowrap;
-  text-align: left;
-  font-weight: normal;
-  font-size: 14px;
-  color: #a0a0a0;
-`;
-const Footer = styled.section`
-  display: flex;
-  flex-direction: row;
-  margin: 32px 16px;
-`;
 
 export default PokemonDetail;
