+++ poke-dex/src/shared/consts/PokemonImageType.ts
