+++ poke-dex/src/shared/ui/PokeMarkChip/PokeMarkChip.styles.ts
@@ -1,14 +1,6 @@
 import styled from "@emotion/styled";
 
-const PokeMarkChip = () => {
-  return (
-    <Chip>
-      <Text>Pokémon</Text>
-    </Chip>
-  );
-}
-
-const Chip = styled.div`
+export const Chip = styled.div`
   display: flex;
   align-items: center;
 
@@ -21,9 +13,7 @@ const Chip = styled.div`
   margin-left: auto;
   margin-right: 16px;
 `;
-const Text = styled.label`
+export const Text = styled.label`
   padding: 0 8px;
   font-size: 14px;
 `;
-
-export default PokeMarkChip;
