+++ poke-dex/src/shared/ui/PokeMarkChip/PokeMarkChip.tsx
@@ -1,0 +1,11 @@
+import { Chip, Text } from './PokeMarkChip.styles'
+
+const PokeMarkChip = () => {
+  return (
+    <Chip>
+      <Text>Pokémon</Text>
+    </Chip>
+  );
+}
+
+export default PokeMarkChip;
