+++ poke-dex/src/shared/ui/PokeNameChip/PokeNameChip.tsx
@@ -1,4 +1,5 @@
-import styled from "@emotion/styled";
+import { Chip, NumberChip, Number, Text } from "./PokeNameChip.styles"
+
 
 interface PokeNameChipProps {
   name: string
@@ -33,28 +34,5 @@ const PokeNameChip = (props: PokeNameChipProps) => {
     </Chip>
   );
 }
-
-const Chip = styled.div`
-  display: flex;
-  align-items: center;
-
-  border: 1px solid #c0c0c0;
-  border-radius: 16px;
-
-  font-weight: bold;
-  box-shadow: 0.5px 0.5px 0 0 #c0c0c0;
-`;
-const NumberChip = styled.div<{ color: string }>`
-  padding: 4px 6px;
-  background-color: ${props => props.color};
-  border-radius: 16px;
-  opacity: 0.8;
-`;
-const Number = styled.label`
-  opacity: 1;
-`;
-const Text = styled.label`
-  margin: 0 8px 0 5px;
-`;
 
 export default PokeNameChip;
