+++ poke-dex/src/widgets/PageHeader/PageHeader.styles.ts
@@ -1,0 +1,20 @@
+import styled from "@emotion/styled";
+
+export const Header = styled.nav`
+    display: flex;
+    padding: 16px 32px;
+    margin-bottom: 16px;
+    border-bottom: 1px solid #c0c0c0;
+`;
+export const Title = styled.h1`
+  margin: 0;
+  font-size: 32px;
+  color: #ffca09;
+  text-shadow: 1px 0 blue, 0 2px blue, 1px 0 blue, 0 1px blue;
+`;
+export const Select = styled.select`
+  display: flex;
+  margin-left: auto;
+  padding: 8px 12px;
+  border-radius: 8px;
+`;
