+++ poke-dex/src/widgets/PageHeader/PageHeader.tsx
@@ -1,10 +1,10 @@
-import styled from "@emotion/styled";
 import { useSelector } from 'react-redux';
 import { Link } from "react-router-dom";
-import { POKEMON_IMAGE_TYPE } from "../Constants";
-import { RootState, useAppDispatch } from "../Store";
+import { POKEMON_IMAGE_TYPE } from "../../shared/consts/PokemonImageType";
+import { RootState, useAppDispatch } from "../../shared/lib/Store";
 import { ChangeEvent } from "react";
-import { PokemonImageKeyType, changeImageType } from "../Store/imageTypeSlice";
+import { PokemonImageKeyType, changeImageType } from "../../shared/lib/Store/imageTypeSlice";
+import { Header, Select, Title } from './PageHeader.styles';
 
 const PageHeader = () => {
   const type = useSelector((state: RootState) => state.imageType.type);
@@ -29,24 +29,5 @@ const PageHeader = () => {
     </Header>
   );
 }
-
-const Header = styled.nav`
-    display: flex;
-    padding: 16px 32px;
-    margin-bottom: 16px;
-    border-bottom: 1px solid #c0c0c0;
-`;
-const Title = styled.h1`
-  margin: 0;
-  font-size: 32px;
-  color: #ffca09;
-  text-shadow: 1px 0 blue, 0 2px blue, 1px 0 blue, 0 1px blue;
-`;
-const Select = styled.select`
-  display: flex;
-  margin-left: auto;
-  padding: 8px 12px;
-  border-radius: 8px;
-`;
 
 export default PageHeader;
